Add support for controlling migrations managed schema
Adds commands parameter (--schema) and configurarion option for
configuring the specific schema to be controlled by migrations
and ignore the changes in other schemas.

The intent is limiting diff results (changelogs) to db objects in the
specified schema, so you can have entities in schemas externally managed
along with app's specific entities.