Fix ConcurrentModificationException when running dbm-gorm-diff
This fix upgrades liquibase-core to version 3.4.2, this has been identified as a
workaround solution to the concurrent modification exception being thrown by
gorm-diff, generate-changelog etc.

Some parts of the api in liquibase changed between 3.3 and 3.4, so while the
original error went away, bringing in 3.4.2 broken the dbmUpdate
functionality. This commit fixes this issue with also and no workarounds
should be needed.

https://github.com/grails-plugins/grails-database-migration/issues/47