Fix missing dependency problems by changing the scope from implementation to compileOnly
The buildsciprt does not resolve dependency version from the Grails BOM, so but defining the dependency scope as implementation add information to the POM without version which breaks when add the plugin to build dependency. Changing the scope to compileOnly will skip dependency from POM which will resolve imissing dependency errors.

Fixes #229